Create the Gitgraphia namespace
The changes I am about to do require me to define something that is
capable of recursively walking backwards from a commit.

I most likely need something I can name and I do not want to chuck it
inside the `Gitgraphia` class.
