+++ bin/gitgraphia
@@ -7,15 +7,15 @@ require 'gitgraphia'
 
 git_sha = 'HEAD'
 
-gitgraphia = Gitgraphia.new
+reader = Gitgraphia::Reader.new
 
-gitgraphia.parents_of(git_sha).each do |parent_sha|
+reader.parents_of(git_sha).each do |parent_sha|
   puts "commit #{git_sha} -[:has_parent]-> commit #{parent_sha}"
 end
 
-tree_sha = gitgraphia.tree_of(git_sha)
+tree_sha = reader.tree_of(git_sha)
 puts "commit #{git_sha} -[:represents]-> tree #{tree_sha}"
 
-gitgraphia.tree_to_files(tree_sha).each do |file|
+reader.tree_to_files(tree_sha).each do |file|
   puts "tree #{tree_sha} -[:contains]-> #{file}"
 end
